Changes tagdetect3
