+++ hk_ros_2021/launch/hk_ros_2021.launch
@@ -1,37 +1,38 @@
 <launch>
 
-    <!-- set /use_sim_time true if replaying bag  (also remember clock flag for rosbag play) --> 
+    <!-- set /use_sim_time true if replaying bag  (also remember clock flag for rosbag play) -->
     <param name="/use_sim_time" value="true"/>
 
     <!-- uncompress camera feed -->
-    <node name="republish" type="republish" pkg="image_transport" output="screen" args="compressed in:=/raspicam_node/image raw out:=/raspicam_node/image" /> 
+    <node name="republish" type="republish" pkg="image_transport" output="screen" args="compressed in:=/raspicam_node/image raw out:=/raspicam_node/image" />
 
     <!-- launch localization (hint: check out the preinstalled "turtlebot3_slam" package) -->
 	<include file="$(find turtlebot3_slam)/launch/turtlebot3_slam.launch">
 		<arg name="open_rviz" value="false"/>
-		
 
 	</include>
-	
 
-	
-	
-    <!-- launch object A detector (april tags) --> 
+
+    <!-- launch object A detector (april tags) -->
 	<include file="$(find apriltag_ros)/launch/continuous_detection.launch">
 		<arg name="camera_name" value="raspicam_node"/>
 		<arg name="image_topic" value="image"/>
 		<arg name="camera_frame" value="camera_rgb_optical_frame"/>
+  </include>
 
-		
-	</include> 
-	
 
     <!-- launch object B detector (geometric shapes) -->
 
+
     <!-- launch object C detector (animals) -->
-	
+	<include file="$(find darknet_ros)/launch/yolo_v3.launch">
+      <arg name="network_param_file"  value="$(find darknet_ros)/config/yolov3.yaml"/>  <!-- Simplified config .yaml file with fewer items-->
+      <arg name="image" value="raspicam_node/image"/>
+  </include>
+
 
     <!-- launch rviz -->
+<<<<<<< HEAD
     <node pkg="rviz" type="rviz" name="rviz" required="true" args="-d $(find hk_ros_2021)/rviz_config/hk_ros_2021.rviz"/>   
     
    
@@ -39,5 +40,12 @@
     <!-- <node pkg="hk_ros_2021" name="tag8" type="odom_tag8_subscriber.py" output="screen"></node> -->
     
     <node pkg="hk_ros_2021" name="Detection_coordinates" type="output_generation_template.py" output="screen"></node> 
+=======
+    <node pkg="rviz" type="rviz" name="rviz" required="true" args="-d $(find hk_ros_2021)/rviz_config/hk_ros_2021.rviz"/>
+
+      <!--<node pkg="hk_ros_2021" name="tag9" type="odom_tag9_subscriber.py" output="screen"></node> -->
+      <!-- <node pkg="hk_ros_2021" name="tag8" type="odom_tag8_subscriber.py" output="screen"></node> -->
+      <!--<node pkg="hk_ros_2021" name="Detection_coordinates" type="output_generation_template.py" output="screen"></node> -->
+>>>>>>> e822fefc6687213372b2955c9332625bcb599905
 
 </launch>
